Rename randomString to randomValue for String, added randomValue properties for Bool and Double
