+++ Tests/AppTests/Extensions/Foundation/Bool/Bool+RandomValue.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension Bool {
+  static var randomValue: Bool {
+    return arc4random_uniform(2) == 0
+  }
+}
