+++ Tests/AppTests/Extensions/Foundation/String/String+RandomValue.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension String {
+  static var randomValue: String {
+    let uuid = UUID().uuidString
+    let upperBound = UInt32(uuid.count - 1)
+    let randomStringLength = Int(arc4random_uniform(upperBound)) + 1
+    let randomStringIndex = String.Index(encodedOffset: randomStringLength)
+    let randString = String(uuid[..<randomStringIndex])
+    return randString
+  }
+}
